requirements.txt
# Core dependencies (required for all)
python-dotenv>=1.0.0
numpy>=1.24.0

# Evaluation (required for metrics_generator.py)
openai-whisper>=20231117

# API Models - install based on what you need:
# openai_api_transcribe.py, groq_whisper_transcribe.py
openai>=1.0.0

# gemini_transcribe.py
google-genai>=1.0.0

# elevenlabs_scribe_transcribe.py
elevenlabs>=1.0.0

# mistral_voxtral_*.py
mistralai>=1.0.0

# mistral_voxtral_mini_transcription_transcribe.py (uses requests directly)
requests>=2.31.0

# azure_foundry_phi4_transcribe.py
azure-ai-inference>=1.0.0
azure-core>=1.28.0

# Local MLX Models (Apple Silicon only):
# mlx_whisper_transcribe.py
# mlx-whisper>=0.4.0

# NVIDIA GPU Models:
# canary_*.py
# torch>=2.0.0
# nemo-toolkit[asr]>=1.20.0
# librosa>=0.10.0
# soundfile>=0.12.0
# pydub>=0.25.0

# Kyutai Models:
# kyutai_stt_*.py
# moshi>=0.1.0
